% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diurnalParams.R
\name{diurnalParams}
\alias{diurnalParams}
\title{Estimate diurnal chlorophyll fluorescence parameters}
\usage{
diurnalParams(filtered.data,FmR.data,save.file,save.path)
}
\arguments{
\item{filtered.data}{the final filtered MONI-PAM data retrieved from cleaning step 6  (i.e., flag6.neighbor in Intro_to_processMONIPAM.Rmd)}

\item{FmR.data}{estimated chlorophyll fluorescence reference data, which can be retrieved from \link{ChlFRef} function in this R package, or prepared by yourself. This data should be a data frame or data table and include columns 'season','head_tree','head','tree_num','FmR' and 'FoR'}

\item{save.file}{TRUE or FALSE. If TRUE, output data will be saved to local folder via save.path argument}

\item{save.path}{local folder for saving the output data generated from this function}
}
\value{
This function will return a data frame including estimated diurnal parameters and original data
}
\description{
This function will estimate diurnal parameters,including rate constant parameters: PQ, NPQ,NPQr; quenching parameters: qLT,qLr, and quantum yield parameters: Phi_NPQ, Phi_NPQr,Phi_NPQs, Phi_fD. Please see calculation method and parameter description in Porcar-Castell et al. 2011 and Zhang et al. XX
}
\references{
reference Porcar‐Castell, A. (2011). A high‐resolution portrait of the annual dynamics of photochemical and non‐photochemical quenching in needles of Pinus sylvestris. Physiologia Plantarum, 143(2), 139-153. https://doi.org/10.1111/j.1399-3054.2011.01488.x
}
