% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter5.lowfm.R
\name{filter5.lowfm}
\alias{filter5.lowfm}
\title{Filter MONI-PAM data step 5,remove additionally abnormal low Fm’ data that is or are adjacent to previously filtered data}
\usage{
filter5.lowfm(moni.data,save.path,save.file,fm1=0.2,fmPhiPratio=3)
}
\arguments{
\item{moni.data}{a data.table or data.frame MONI-PAM data generated from \link{filter4.maxYield} function.}

\item{save.path}{local folder for saving your output file}

\item{save.file}{If this argument is set as TRUE, the returned file will be saved to local folder, if FALSE, the file will not be saved into local folder}

\item{fm1}{the threshold of percentage change of Fm' between time2 and time1. Default value is 0.2, we recommend this argument can be adjusted from 0.05 to 0.5 by an interval of 0.05.}

\item{fmPhiPratio}{the threshold of ratio between percentage change of Fm' between time2 and time1 and of Yield between time2 and time1. Default value is 3, we recommend this argument can be adjust between 2 and 5 by an interval of 1.}
}
\value{
\link{filter5.lowfm} will return a data table. Meanwhile, if save.file = TRUE, the output data.table will also be saved into local folder as a 'MONI_Year1_Year2_filter5lowfm.dat' file, where Year1 and Year2 are the minimum and maximum year during this observation season respectively. This output file will contain one new column compared with output file from \link{filter4.maxYield} function named as 'flag5.lowfm'. This column only contain two values: 0 and 1, where 0 means F', Fm' and Yield in corresponding row(s) are abnormal data and should be removed from the dataset and 1 means good dataset.
}
\description{
Details see Zhang et al.,202X. paper link url.
}
