% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter6.neighbor.R
\name{filter6.neighbor}
\alias{filter6.neighbor}
\title{Filter MONI-PAM data step 6,remove additionally a few more data points adjacent to the previously filtered data}
\usage{
filter6.neighbor(moni.data,save.path,save.file,expand.time)
}
\arguments{
\item{moni.data}{a data.table or data.frame MONI-PAM data generated from \link{filter5.lowfm} function.}

\item{save.path}{local folder for saving your output file}

\item{save.file}{If this argument is set as TRUE, the returned file will be saved to local folder, if FALSE, the file will not be saved into local folder}

\item{expand.time}{an integer value.  This value means a time window in minutes and allows users using different sampling intervals to decide how many points adjacent to filtered data they want to remove. For example, for a 20-min MONI-PAM measurement interval dataset, if expand.time=60 (it means 60 mins), and two points of both left and right side of current removed point will be removed from this function.}
}
\value{
\link{filter6.neighbor} will return a data table. Meanwhile, if save.file = TRUE, the output data.table will also be saved into local folder as a 'MONI_Year1_Year2_filter6passTOneighbor.dat' file, where Year1 and Year2 are the minimum and maximum year during this observation season respectively. This output file will contain one new column compared with output file from \link{filter5.lowfm} function named as 'flag6.neighbor'. This column only contain two values: 0 and 1, where 0 means F', Fm' and Yield in corresponding row(s) are abnormal data and should be removed from the dataset and 1 means good dataset.
}
\description{
Details see Zhang et al.,202X. paper link url.
}
