% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatMONIdata.R
\name{formatMONIdata}
\alias{formatMONIdata}
\title{convert MONIPAM data variables to correct data format}
\usage{
formatMONIdata(moni.data)
}
\arguments{
\item{moni.data}{a combined organized MONI-PAM data which is generated from \link{readMONIPAM} function, or a MONI-PAM data after data filtering which is generated from filter function in this package such as \link{filter1.NA}, \link{filter2.night} and so on.}
}
\value{
\link{formatMONIdata} will generate a MONI-PAM data with all the variables in proper data format for later data filtering and processing
}
\description{
This function will convert the MONI-PAM variables to proper data format for easily used for data filtering and processing
}
