% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.checkfilter.R
\name{plot.checkfilter}
\alias{plot.checkfilter}
\title{MONI-PAM data visualization with data before and after filter}
\usage{
\method{plot}{checkfilter}(newdata, rawdata, save.path, diary.data)
}
\arguments{
\item{newdata}{one of MONI-PAM data after data filter (e.g., output file from \link{filter3.day} function)}

\item{rawdata}{original organized combined MONI-PAM data which is generated after read.MONIPAM function or other previous filtered data such as output data from filter2.night function.}

\item{save.path}{local folder for saving the plotted figures generated from this function}

\item{diary.data}{this is optional, the diary file which can be such as .csv or .dat file but with a specific data format. You can check the example data from this package using data("dairy_for_plot"). In this dataset, head and tree_num should be same with your MONI-PAM data, text in 'remark.plot' column will be shown in the plotted figure to show what has happened for that 'head'/ 'tree_num'/'head_tree' on which 'datetime' so it should be very short and clean. 'remark' column can can contain full description for remark.plot. 'manager' column shows who found that issue or conducted certain activity for that sensor (e.g., calibration)}
}
\value{
\link{plot.checkfilter} will not return/show the plotted figures in the end, instead, it will save the plotted figures to your folder directly.
}
\description{
This function will visualize two MONI-PAM data, one data before filter (rawdata) and another data after filter (newdata) (including F',FM',Yield,PAR, and temperature in one figure) with 10 days as a group plotting for each head and the dairy note will be marked in the figure if the note file is available. At the same time, this function will also show the sunlight time (colorful vertical lines) in the plot, sunrise and sunsetStart are 'indianred', solarNoon is 'goldenrod', and dawn and dusk are 'purple'. In the plotted figure, red, pink, and orange point means the filtered or flagged Yield, Fm', and F'data respectively in the newdata compared with rawdata.
}
