% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMONIdiary.R
\name{plotMONIdiary}
\alias{plotMONIdiary}
\title{MONI-PAM data visualization with dairy note marked in the figure}
\usage{
plotMONIdiary(moni.data,diary.data,plot.title,save.path)
}
\arguments{
\item{moni.data}{a combined organized MONI-PAM data which is generated from \link{readMONIPAM} function, or a MONI-PAM data after data filtering which is generated from filter function in this package such as \link{filter1.NA}, \link{filter2.night} and so on.}

\item{diary}{TRUE or FALSE. If diary = TRUE, dairy data should be assigned.}

\item{diary.data}{a diary file which can be such as .csv or .dat file but with a specific data format. You can check the example data from this package using data("dairy_for_plot"). In this dataset, head and tree_num should be same with your MONI-PAM data, text in 'remark.plot' column will be shown in the plotted figure to show what has happened for that 'head'/ 'tree_num'/'head_tree' on which 'datetime' so it should be very short and clean. 'remark' column can can contain full description for remark.plot. 'manager' column shows who found that issue or conducted certain activity for that sensor (e.g., calibration)}

\item{plot.title}{any text to describe clearly about your figure content}

\item{save.path}{local folder for saving your plotted figures}
}
\value{
\link{plotMONIdiary} will not return/show the plotted figures in the end, instead, it will save the plotted figures to your folder directly.
}
\description{
This function will visualize MONI-PAM data (F',FM',Yield,PAR, and temperature in one figure) with 10 days as a group plotting for each head and the dairy note will be marked in the figure if the note file is available.
}
