% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSeasonPara.R
\name{plotSeasonPara}
\alias{plotSeasonPara}
\title{MONI-PAM FV/FM data visualization}
\usage{
plotSeasonPara(season.param,save.path,diary,diary.data)
}
\arguments{
\item{season.param}{seasonal chlorophyll fluorescence parameters retrieved from \link{seasonalParams} function in this R package}

\item{save.path}{local folder for saving the plotted figures generated from this function}

\item{diary}{TRUE or FALSE. If diary = TRUE, dairy data should be assigned.}

\item{diary.data}{a diary file which can be such as .csv or .dat file but with a specific data format. You can check the example data from this package using data("diary_2014_2015"). In this dataset, head and tree_num should be same with your MONI-PAM data, text in 'remark.plot' column will be shown in the plotted figure to show what has happened for that 'head'/ 'tree_num'/'head_tree' on which 'datetime' so it should be very short and clean. 'remark' column can can contain full description for remark.plot. 'manager' column shows who found that issue or conducted certain activity for that sensor (e.g., calibration)}
}
\value{
\link{plotSeasonPara} will not return/show the plotted figures in the end, instead, it will save the plotted figures to your folder directly.
}
\description{
This function will visualize MONI-PAM FV/FM data including F0,FM,FV/FM,PAR, and temperature in one figure each head.
}
