% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMONIPAM.R
\name{readMONIPAM}
\alias{readMONIPAM}
\title{read and organize  MONI-PAM data}
\usage{
read.MONIPAM(source.path,pam.pattern,save.path,site.lat,site.lon,local.tz,tz.summer,tz.winter,measure.time)
}
\arguments{
\item{source.path}{local path where original MONI-PAM data are stored}

\item{pam.pattern}{MONI-PAM data extension, '.pam' or '.PAM'}

\item{save.path}{local path where the exported data will be saved}

\item{site.lat}{latitude of the study site}

\item{site.lon}{longitude of the study site}

\item{local.tz}{time zone for the study site, for example, it is 'EET' for Finland}

\item{tz.summer}{UTC offset of study site in summer, for example, it is 3 for Finland}

\item{tz.winter}{UTC offset of study site in winter, for example, it is 2 for Finland}

\item{measure.time}{which time zone used for measurement: 'winter','summer' or 'local'}
}
\value{
list 1: combined original data; list 2: combined organized data and extra columns for further data cleaning purpose
}
\description{
The output of this function will include many files that will be saved into
local folder, including
(1) rawEachFilename_Year.csv: organized each single file.
(2) four information files: rawFile_noHeadInfo.txt, rawFile_newDeviceAdded.txt,
rawFile_noTreeDefined.txt, and rawFile_noDataRecord.txt. These files will be
saved into local folder if the according condition occurs, then you need
to mordify your MONI-PAM files based on error or warning message in
order to continue smoothly organize your data and further data cleaning.
(3) "preprocesMONI_head_timerange_Year1_Year2.dat": a summary of head,
tree-number and measuring time information.
(4) "preprocesMONI_Year1_Year2.dat": combined all the
organized data and this is the file used for further data
processing and cleaning. In this file, columns "plot.group",
"dateBack12h",datetimeBack12h","sunrise","sunriseEnd","solarNoon",
"sunsetStart", "sunset", and "dusk" were added for later easily
data cleaning and data visualization purpose.
}
