% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.FvFm.R
\name{scale.fvfm}
\alias{scale.fvfm}
\title{obtain extreme values of MONI-PAM Fv/Fm data}
\usage{
\method{scale}{fvfm}(moni.fvfm)
}
\arguments{
\item{moni.fvfm}{MONI-PAM fvfm data retrieved from \link{FindFvFm} function in this R package}
}
\value{
\link{scale.fvfm} will return a data.table which contains the maximum and minimum values of F0,FM, Fv/Fm, PAR, and temperature values for each MONI-head respectively.
}
\description{
These extreme values of MONI-PAM data will be used for MONI-PAM data visualization
}
