% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.monidata.R
\name{scale.monidata}
\alias{scale.monidata}
\title{obtain extreme values of MONI-PAM data}
\usage{
\method{scale}{monidata}(moni.data)
}
\arguments{
\item{moni.data}{a combined organized MONI-PAM data which is generated from \link{read.MONIPAM} function, or a MONI-PAM data after data filtering which is generated from filter function in this package such as \link{filter1.NA}, \link{filter2.night} and so on.}
}
\value{
\link{scale.monidata} will return a data.table which contains the maximum and minimum values of F',FM',Yield, PAR, and temperature values for each plotting groups (every 10 days as a group) and each MONI-head
}
\description{
These extreme values of MONI-PAM data will be used for MONI-PAM data visualization
}
