% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFvFm.R
\name{scaleFvFm}
\alias{scaleFvFm}
\title{obtain extreme values of MONI-PAM Fv/Fm data}
\usage{
scaleFvFm(moni.fvfm)
}
\arguments{
\item{moni.fvfm}{MONI-PAM fvfm data retrieved from \link{FindFvFm} function in this R package}
}
\value{
\link{scaleFvFm} will return a data.table which contains the maximum and minimum values of F0,FM, Fv/Fm, PAR, and temperature values for each MONI-head respectively.
}
\description{
These extreme values of MONI-PAM data will be used for MONI-PAM data visualization
}
