% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonalParams.R
\name{seasonalParams}
\alias{seasonalParams}
\title{Estimate seasonal chlorophyll fluorescence parameters}
\usage{
seasonalParams(filtered.fvfm,FmR.data,save.file,save.path)
}
\arguments{
\item{save.file}{TRUE or FALSE. If TRUE, output data will be saved to local folder via save.path argument}

\item{save.path}{local folder for saving the plotted figures generated from this function}

\item{raw.fvfm}{MONI-PAM fvfm data retrieved from original data (i.e., flag0.fvfm in Intro_to_processMONIPAM.Rmd) by using \link{FindFvFm} function in this R package}

\item{filter.fvfm}{the final MONI-PAM fvfm data retrieved from cleaning step 6 (i.e., flag6.fvfm in Intro_to_processMONIPAM.Rmd) by using \link{FindFvFm} function in this R package}
}
\value{
This function will return a data frame including new estimated seasonal parameters and Fv/Fm data
}
\description{
This function will estimate seasonal parameters,including PQs, qLs, and NPQs. Please see calculation method and parameter description in Porcar-Castell et al. 2011 and Zhang et al. XX.
}
\references{
reference Porcar‐Castell, A. (2011). A high‐resolution portrait of the annual dynamics of photochemical and non‐photochemical quenching in needles of Pinus sylvestris. Physiologia Plantarum, 143(2), 139-153. https://doi.org/10.1111/j.1399-3054.2011.01488.x
}
